New format-Bundle/Olm classes
