+++ .jenkins/Jenkinsfile-pr
@@ -181,7 +181,7 @@ pipeline {
             steps {
                 script {
                     catchError {
-                        // For upgrade tests running with locally built images we need to change tag in StrimziUpgradeST.yaml
+                        // For upgrade tests running with locally built images we need to change tag in BundleUpgrade.yaml
                         if ("${env.TEST_PROFILE}".contains("upgrade")) {
                             lib.prepareUpgradeSTs(env.WORKSPACE, env.DOCKER_REGISTRY, env.DOCKER_TAG)
                         }
