+++ .jenkins/jenkins.groovy
@@ -40,11 +40,11 @@ def buildStrimziImages() {
 def prepareUpgradeSTs(String workspace, String dockerRegistry, String dockerTag) {
     println("[INFO] Update files for upgrade procedure")
     sh(script: """
-        sed -i 's#:latest#:${dockerTag}#g' ${workspace}/systemtest/src/test/resources/upgrade/StrimziUpgradeST.yaml ${workspace}/install/cluster-operator/060-Deployment-strimzi-cluster-operator.yaml
+        sed -i 's#:latest#:${dockerTag}#g' ${workspace}/systemtest/src/test/resources/upgrade/BundleUpgrade.yaml ${workspace}/install/cluster-operator/060-Deployment-strimzi-cluster-operator.yaml
         sed -i 's#quay.io/strimzi/#${dockerRegistry}/strimzi/#g' ${workspace}/install/cluster-operator/060-Deployment-strimzi-cluster-operator.yaml
         sed -i 's#/opt/${dockerRegistry}#/opt#g' ${workspace}/install/cluster-operator/060-Deployment-strimzi-cluster-operator.yaml
     """)
-    sh(script: "cat ${workspace}/systemtest/src/test/resources/upgrade/StrimziUpgradeST.yaml")
+    sh(script: "cat ${workspace}/systemtest/src/test/resources/upgrade/BundleUpgrade.yaml")
     sh(script: "cat ${workspace}/install/cluster-operator/060-Deployment-strimzi-cluster-operator.yaml")
 }
