+++ systemtest/src/main/java/io/strimzi/systemtest/resources/operator/configuration/OlmConfiguration.java
@@ -19,7 +19,6 @@ public class OlmConfiguration {
     private String namespaceName;
     private String featureGates = Environment.STRIMZI_FEATURE_GATES;
     private String olmAppBundlePrefix = Environment.OLM_APP_BUNDLE_PREFIX;
-    private String olmOperatorDeploymentName = Environment.OLM_OPERATOR_DEPLOYMENT_NAME;
     private String olmOperatorName = Environment.OLM_OPERATOR_NAME;
     private String olmSourceName = Environment.OLM_SOURCE_NAME;
     private String olmSourceNamespace = Environment.OLM_SOURCE_NAMESPACE;
@@ -111,7 +110,7 @@ public class OlmConfiguration {
     }
 
     public String getOlmOperatorDeploymentName() {
-        return olmOperatorDeploymentName;
+        return olmAppBundlePrefix + "-v" + operatorVersion;
     }
 
     public String getOlmOperatorName() {
