+++ systemtest/src/test/java/io/strimzi/systemtest/upgrade/OlmUpgradeIsolatedST.java
@@ -19,7 +19,6 @@ import io.strimzi.systemtest.storage.TestStorage;
 import io.strimzi.systemtest.utils.ClientUtils;
 import io.strimzi.systemtest.utils.FileUtils;
 import io.strimzi.systemtest.utils.RollingUpdateUtils;
-import io.strimzi.systemtest.utils.TestKafkaVersion;
 import io.strimzi.systemtest.utils.kubeUtils.controllers.DeploymentUtils;
 import io.strimzi.systemtest.annotations.IsolatedSuite;
 import io.strimzi.test.k8s.KubeClusterResource;
@@ -29,6 +28,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtensionContext;
+import io.strimzi.systemtest.upgrade.VersionModificationDataLoader.ModificationType;
 
 import java.io.File;
 import java.io.IOException;
@@ -57,40 +57,26 @@ public class OlmUpgradeIsolatedST extends AbstractUpgradeST {
     @Test
     void testStrimziUpgrade(ExtensionContext extensionContext) throws IOException {
         // Fetch upgrade data from yaml and parse them to class
-        UpgradeDowngradeData latestUpgradeData = new UpgradeDowngradeDatalist().getLatestUpgradeData();
-        TestKafkaVersion testKafkaVersion = TestKafkaVersion.getLatestSupportedKafkaVersion();
-
-        // Create and set procedures -> replace kafka version after operator upgrade.
-        // This is needed, because for upgrade it's generated by method which generated data for parametrized test
-        latestUpgradeData.setProcedures(new UpgradeKafkaVersion(testKafkaVersion.version()));
-
-        // Set OLM operator version which we want to upgrade to
-        latestUpgradeData.setToVersion(Environment.OLM_OPERATOR_LATEST_RELEASE_VERSION);
-
-        // Perform upgrade and verify results
-        performUpgradeVerification(latestUpgradeData, extensionContext);
-    }
-
-    private void performUpgradeVerification(UpgradeDowngradeData upgradeData, ExtensionContext extensionContext) throws IOException {
+        OlmVersionModificationData olmUpgradeData = new VersionModificationDataLoader(ModificationType.OLM_UPGRADE).getOlmUpgradeData();
         final TestStorage testStorage = new TestStorage(extensionContext);
-        final String toVersion = upgradeData.getToVersion();
-        final String fromVersion = upgradeData.getFromVersion();
+        final String toVersion = olmUpgradeData.getToVersion();
+        final String fromVersion = olmUpgradeData.getFromVersion();
 
         LOGGER.info("====================================================================================");
         LOGGER.info("---------------- Updating Cluster Operator version " + fromVersion + "=>" + toVersion + " -----------------");
         LOGGER.info("====================================================================================");
         LOGGER.info("-------------------------- Upgrade data used in this test --------------------------");
-        LOGGER.info(upgradeData.toString());
+        LOGGER.info(olmUpgradeData.toString());
         LOGGER.info("====================================================================================");
 
         // Install operator via Olm
         //  1. Create subscription with manual approval and operator group if needed
         //  2. Approve installation (get install-plan and patch it)
-        clusterOperator.runManualOlmInstallation(upgradeData.getFromOlmChannelName());
+        clusterOperator.runManualOlmInstallation(olmUpgradeData.getFromOlmChannelName(), fromVersion);
 
         // In this test we intend to setup Kafka once at the beginning and then upgrade it with CO
-        File dir = FileUtils.downloadAndUnzip(upgradeData.getUrlFrom());
-        File kafkaYaml = new File(dir, upgradeData.getFromExamples() + "/examples/kafka/kafka-persistent.yaml");
+        File dir = FileUtils.downloadAndUnzip(olmUpgradeData.getFromUrl());
+        File kafkaYaml = new File(dir, olmUpgradeData.getFromExamples() + "/examples/kafka/kafka-persistent.yaml");
 
         LOGGER.info("Deploy Kafka from file: {}", kafkaYaml.getPath());
         KubeClusterResource.cmdKubeClient().create(kafkaYaml);
@@ -101,15 +87,15 @@ public class OlmUpgradeIsolatedST extends AbstractUpgradeST {
         HashMap<String, Object> topicConfig = new HashMap<String, Object>();
         topicConfig.put("min.insync.replicas", 2);
 
-        KafkaTopic kafkaUpgradeTopic = new YAMLMapper().readValue(new File(dir, upgradeData.getFromExamples() + "/examples/topic/kafka-topic.yaml"), KafkaTopic.class);
+        KafkaTopic kafkaUpgradeTopic = new YAMLMapper().readValue(new File(dir, olmUpgradeData.getFromExamples() + "/examples/topic/kafka-topic.yaml"), KafkaTopic.class);
         kafkaUpgradeTopic = new KafkaTopicBuilder(kafkaUpgradeTopic)
             .editMetadata()
             .withNamespace(clusterOperator.getDeploymentNamespace())
-                .withName(topicUpgradeName)
+            .withName(topicUpgradeName)
             .endMetadata()
             .editSpec()
-                .withReplicas(3)
-                .addToConfig(topicConfig)
+            .withReplicas(3)
+            .addToConfig(topicConfig)
             .endSpec()
             .build();
 
@@ -126,7 +112,6 @@ public class OlmUpgradeIsolatedST extends AbstractUpgradeST {
             .build();
 
         resourceManager.createResource(extensionContext, kafkaBasicClientJob.producerStrimzi(), kafkaBasicClientJob.consumerStrimzi());
-        ClientUtils.waitForClientsSuccess(testStorage.getProducerName(), testStorage.getConsumerName(), clusterOperator.getDeploymentNamespace(), testStorage.getMessageCount());
 
         String clusterOperatorDeploymentName = ResourceManager.kubeClient().getDeploymentNameByPrefix(Environment.OLM_OPERATOR_DEPLOYMENT_NAME);
         LOGGER.info("Old deployment name of Cluster Operator is {}", clusterOperatorDeploymentName);
@@ -139,10 +124,10 @@ public class OlmUpgradeIsolatedST extends AbstractUpgradeST {
             .withOperatorVersion(toVersion)
             .build();
 
-        clusterOperator.getOlmResource().updateSubscription(upgradeOlmConfig);
+        clusterOperator.updateSubscription(upgradeOlmConfig);
 
         // Cluster Operator upgrade
-        clusterOperator.getOlmResource().upgradeClusterOperator(clusterOperator.getDeploymentNamespace(), upgradeOlmConfig.getCsvName());
+        clusterOperator.upgradeClusterOperator(upgradeOlmConfig);
 
         clusterOperatorDeploymentName = ResourceManager.kubeClient().getDeploymentNameByPrefix(Environment.OLM_OPERATOR_DEPLOYMENT_NAME);
         LOGGER.info("New deployment name of Cluster Operator is {}", clusterOperatorDeploymentName);
@@ -160,12 +145,11 @@ public class OlmUpgradeIsolatedST extends AbstractUpgradeST {
 
         // ======== Kafka upgrade starts ========
         logPodImages(clusterName);
-        changeKafkaAndLogFormatVersion(upgradeData, extensionContext);
+        changeKafkaAndLogFormatVersion(olmUpgradeData, extensionContext);
         logPodImages(clusterName);
         // ======== Kafka upgrade ends ========
 
-        // Send and wait for messages
-        resourceManager.createResource(extensionContext, kafkaBasicClientJob.producerStrimzi(), kafkaBasicClientJob.consumerStrimzi());
+        // Wait for messages of previously created clients
         ClientUtils.waitForClientsSuccess(testStorage.getProducerName(), testStorage.getConsumerName(), clusterOperator.getDeploymentNamespace(), testStorage.getMessageCount());
 
         // Check for errors in Cluster Operator log
