+++ systemtest/src/test/resources/upgrade/BundleUpgrade.yaml
@@ -1,12 +1,11 @@
 #  --- Structure ---
-#- fromVersion: String - From this version we will upgrade the CO to latest = "HEAD"
-#  fromExamples: String - Folder name which contains examples to deploy. Downloaded from "urlFrom"
+#  fromVersion: String - From this version we will upgrade the CO to latest = "HEAD"
+#  fromExamples: String - Folder name which contains examples to deploy. Downloaded from "fromUrl"
 #  oldestKafka: String - version of Kafka, used in initial deployment of Kafka CR
-#  urlFrom: String - url, from which examples are downloaded
+#  fromUrl: String - url, from which examples are downloaded
 #  fromKafkaVersionsUrl: String - url, from which is kafka-versions.yaml file downloaded
-#  fromOlmChannelName: String - name of the olm channel we want to start upgrade from
 #  conversionTool:
-#    urlToConversionTool:  String - url, from which conversion tool is downloaded
+#    toUrlConversionTool:  String - url, from which conversion tool is downloaded
 #    toConversionTool: String - name of folder, which contains conversion tool
 #  additionalTopics: int - number of additional topics, which will be created
 #  imagesAfterOperations:  Strings - images that corresponding pods should contain. They are checked after upgrade
@@ -25,12 +24,11 @@
 #  featureGatesAfter: String - FG added to `STRIMZI_FEATURE_GATES` environment variable, on upgrade of CO
 #  --- Structure ---
 
-- fromVersion: 0.32.0
-  fromExamples: strimzi-0.32.0
+- fromVersion: 0.33.2
+  fromExamples: strimzi-0.33.2
   oldestKafka: 3.2.0
-  urlFrom: https://github.com/strimzi/strimzi-kafka-operator/releases/download/0.32.0/strimzi-0.32.0.zip
-  fromKafkaVersionsUrl: https://raw.githubusercontent.com/strimzi/strimzi-kafka-operator/0.32.0/kafka-versions.yaml
-  fromOlmChannelName: strimzi-0.32.x
+  fromUrl: https://github.com/strimzi/strimzi-kafka-operator/releases/download/0.33.2/strimzi-0.33.2.zip
+  fromKafkaVersionsUrl: https://raw.githubusercontent.com/strimzi/strimzi-kafka-operator/0.33.2/kafka-versions.yaml
   additionalTopics: 2
   imagesAfterOperations:
     zookeeper: strimzi/kafka:latest-kafka-3.4.0
@@ -46,12 +44,11 @@
     reason: Test is working on environment, where k8s server version is < 1.22
   featureGatesBefore: ""
   featureGatesAfter: ""
-- fromVersion: 0.32.0
-  fromExamples: strimzi-0.32.0
+- fromVersion: 0.33.2
+  fromExamples: strimzi-0.33.2
   oldestKafka: 3.2.0
-  urlFrom: https://github.com/strimzi/strimzi-kafka-operator/releases/download/0.32.0/strimzi-0.32.0.zip
-  fromKafkaVersionsUrl: https://raw.githubusercontent.com/strimzi/strimzi-kafka-operator/0.32.0/kafka-versions.yaml
-  fromOlmChannelName: strimzi-0.32.x
+  fromUrl: https://github.com/strimzi/strimzi-kafka-operator/releases/download/0.33.2/strimzi-0.33.2.zip
+  fromKafkaVersionsUrl: https://raw.githubusercontent.com/strimzi/strimzi-kafka-operator/0.33.2/kafka-versions.yaml
   additionalTopics: 2
   imagesAfterOperations:
     zookeeper: strimzi/kafka:latest-kafka-3.4.0
