+++ systemtest/src/test/resources/upgrade/OlmUpgrade.yaml
@@ -1,0 +1,13 @@
+#  --- Structure ---
+#  fromVersion: String - version from which we start olm upgrade to latest released olm operator version (in constants)
+#  fromOlmChannelName: String - (Olm specific) - name of the olm channel from which we start olm upgrade to "stable" channel
+#  fromExamples: String - Folder name which contains examples to deploy. Downloaded from "fromUrl"
+#  fromUrl: String - url, from which examples are downloaded
+#  fromKafkaVersionsUrl: String - url, from which is kafka-versions.yaml file downloaded
+#  --- Structure ---
+
+fromVersion: 0.32.0
+fromOlmChannelName: strimzi-0.32.x
+fromExamples: strimzi-0.32.0
+fromUrl: https://github.com/strimzi/strimzi-kafka-operator/releases/download/0.32.0/strimzi-0.32.0.zip
+fromKafkaVersionsUrl: https://raw.githubusercontent.com/strimzi/strimzi-kafka-operator/0.32.0/kafka-versions.yaml
